# packages/contracts/openapi/v1.yaml
...
openapi: 3.1.0
info:
  description: |-
    Version 1 of the Chaya 02 HTTP API. Breaking changes require a new /api/v2 document.

    Generated from the backend's controllers; packages/contracts/openapi/v1.yaml is the committed output and OpenApiContractTest fails when it drifts.

    Roles (x-chaya-roles) under /api/v1/venues/{venueId}/ are the roles the caller holds at that venue. A resource outside the caller's organization or venue answers 404, never 403.

    Errors are RFC 9457 problems with a stable machine-readable `code`.
  title: Chaya 02 API
  version: v1
servers:
- description: The API's own origin
  url: /
paths:
...
  /api/v1/venues/{venueId}/floors/{floorId}/reconstructions:
    get:
      operationId: reconstructionList
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: floorId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReconstructionVersion"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      - viewerToken: []
      summary: "The floor's reconstruction versions, newest first"
      tags:
      - reconstruction-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
      - PUBLIC_VIEWER
  /api/v1/venues/{venueId}/floors/{floorId}/reconstructions/latest:
    get:
      operationId: reconstructionLatest
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: floorId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reconstruction"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      - viewerToken: []
      summary: The floor's current reconstruction (its current FINALIZED scan version)
      tags:
      - reconstruction-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
      - PUBLIC_VIEWER
...
  /api/v1/venues/{venueId}/reconstructions/{runId}:
    get:
      operationId: reconstructionGet
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: runId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reconstruction"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      - viewerToken: []
      summary: One reconstruction run and its artifacts
      tags:
      - reconstruction-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
      - PUBLIC_VIEWER
  /api/v1/venues/{venueId}/reconstructions/{runId}/artifacts/{kind}:
    get:
      description: "Through the backend, not a presigned URL, so venue/org scope and\
        \ the contains_pii check apply to every read. Bytes are verified against the\
        \ registered SHA-256 while streaming."
      operationId: reconstructionArtifact
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: runId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: kind
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
                format: binary
          description: The artifact's bytes; Content-Type is its registered type
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "BAD_REQUEST: unknown viewer artifact kind"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      - viewerToken: []
      summary: Stream one artifact of a reconstruction run
      tags:
      - reconstruction-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
      - PUBLIC_VIEWER
...
  /api/v1/venues/{venueId}/scan-versions/{scanVersionId}/artifacts/{kind}:
    get:
      description: Bytes are verified against the registered SHA-256 while streaming;
        a mismatch ends the response early.
      operationId: reconstructionVersionArtifact
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: scanVersionId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: kind
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
                format: binary
          description: The artifact's bytes; Content-Type is its registered type
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "BAD_REQUEST: unknown viewer artifact kind"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      - viewerToken: []
      summary: Stream one artifact pinned by a scan version
      tags:
      - reconstruction-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
      - PUBLIC_VIEWER
  /api/v1/venues/{venueId}/scan-versions/{scanVersionId}/reconstruction:
    get:
      description: Exactly its pinned artifacts and its recorded coordinate frame.
      operationId: reconstructionVersion
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: scanVersionId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reconstruction"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: VERSION_NOT_FINALIZED
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      - viewerToken: []
      summary: A FINALIZED scan version's reconstruction
      tags:
      - reconstruction-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
      - PUBLIC_VIEWER
...
components:
  schemas:
...
    ArtifactRef:
      type: object
      properties:
        contentType:
          type: string
        kind:
          type: string
        sha256:
          type: string
        sizeBytes:
          type: integer
          format: int64
        url:
          type: string
...
    FrameView:
      type: object
      properties:
        calibratedAt:
          type: string
          format: date-time
        calibratedBy:
          type: string
        canonical:
          type: boolean
        controlPointRmsM:
          type: number
          format: double
        floorId:
          type: string
          format: uuid
        gravityDisagreementDeg:
          type: number
          format: double
        gravitySource:
          type: string
        gravityStatus:
          type: string
        gravityUpReconstruction:
          type: array
          items:
            type: number
            format: double
        horizontalDatum:
          type: string
        id:
          type: string
          format: uuid
        inputs:
          type: object
          additionalProperties: {}
        method:
          type: string
        metricStatus:
          type: string
        residuals:
          type: object
          additionalProperties: {}
        rotation:
          $ref: "#/components/schemas/Rotation"
        scale:
          type: number
          format: double
        scaleRelativeSpread:
          type: number
          format: double
        scaleSource:
          type: string
        sourceRunId:
          type: string
          format: uuid
        status:
          type: string
        translation:
          $ref: "#/components/schemas/Vec3"
        units:
          type: string
        upAxis:
          type: string
        version:
          type: integer
          format: int32
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        code:
          type: string
          description: "Stable machine-readable error code, e.g. NOT_FOUND, BAD_REQUEST,\
            \ VERSION_INCOMPLETE. Absent on Spring's own request-parsing and validation\
            \ problems."
        detail:
          type: string
        instance:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
          example: about:blank
      required:
      - status
...
    Reconstruction:
      type: object
      properties:
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactRef"
        coordinateFrame:
          $ref: "#/components/schemas/FrameView"
        floorId:
          type: string
          format: uuid
        generatedAt:
          type: string
          format: date-time
        parentVersionId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        runQuality:
          type: string
        runStatus:
          type: string
        scanId:
          type: string
          format: uuid
        scanVersionId:
          type: string
          format: uuid
        versionNumber:
          type: integer
          format: int32
    ReconstructionVersion:
      type: object
      properties:
        current:
          type: boolean
        floorId:
          type: string
          format: uuid
        generatedAt:
          type: string
          format: date-time
        parentVersionId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        runQuality:
          type: string
        runStatus:
          type: string
        scanVersionId:
          type: string
          format: uuid
        versionNumber:
          type: integer
          format: int32
...
    Rotation:
      type: object
      properties:
        w:
          type: number
          format: double
        x:
          type: number
          format: double
        "y":
          type: number
          format: double
        z:
          type: number
          format: double
...
    Vec3:
      type: object
      properties:
        x:
          type: number
          format: double
        "y":
          type: number
          format: double
        z:
          type: number
          format: double
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "Keycloak access token for the chaya-api audience. Organization,\
        \ roles and venue memberships come from its claims."
      scheme: bearer
      type: http
    viewerToken:
      description: "Short-lived public-viewer token from POST /api/v1/public/viewer-token,\
        \ bound to one venue; acts as PUBLIC_VIEWER. Never together with a bearer\
        \ token (400)."
      in: header
      name: X-Chaya-Viewer-Token
      type: apiKey